+++ horaservices/src/components/wonderland/NoteSkeleton/index.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import "./NoteSkeleton.css";
+
+const NoteSkeleton = () => {
+  return <div className="note-skeleton-box"></div>;
+};
+
+export default NoteSkeleton;
